Redesign aggregates
