+++ Hackathon/Team3_Aggregation/README.MD
@@ -11,9 +11,8 @@ Aggregat 1.2: Name Zentrum, PLZ Zentrum, Geschlecht, Anzahl
 Aggregat 1.3: Name Zentrum, PLZ Zentrum, Entfernungskategorie, Anzahl
 Aggregat 1.4: Name Zentrum, PLZ Zentrum, Altersgruppe, Geschlecht, Anzahl
 
-Aggregat 2: Fokus auf einzelne Diagnosen
+* Aggregatgruppe 2: Fokus auf einzelne Diagnosen
+Wie Aggregate 1, aber ergänzt um Diagnose (vergröbert, z.B. 3-Steller oder Gruppe)
 
-* Wie Aggregate 1, aber ergänzt um Diagnose (vergröbert, z.B. 3-Steller oder Gruppe)
-Aggregat 3: Fokus auf Wohnort & Diagnose
-
-* Wie Aggregate 2, aber ergänzt um Bundesland des Wohnorts
+* Aggregatgruppe 3: Fokus auf Wohnort & Diagnose
+Wie Aggregate 2, aber ergänzt um Bundesland des Wohnorts
