+++ Hackathon/Team3_Aggregation/config/agg-2.1.csv
@@ -1,12 +1,1 @@
-Name des Zentrums;PLZ des Zentrums;Diagnose;Anzahl
-Charité;10117;E84;6
-Charité;10117;R65;5
-Charité;10117;D86;6
-Charité;10117;Q05;11
-Charité;10117;M09;10
-Charité;10117;D18;6
-Charité;10117;U07;5
-Charité;10117;Q91;7
-Charité;10117;G35;8
-Charité;10117;D83;5
-Charité;10117;Q96;7
+Name des Zentrums;PLZ des Zentrums;Alter;Diagnose1;Diagnose2;Anzahl
