--- Hackathon/Team3_Aggregation/config/agg-3.1.csv
@@ -1,1 +1,0 @@
-Name des Zentrums;PLZ des Zentrums;Alter;Diagnose1;Diagnose2;Anzahl
